+++ bin/templates/components.tsx
@@ -4,7 +4,7 @@
 ${components.map(component => "import " + component.capitalized + " from '" + component.root + component.name + "'").join('\n')}
 
 import branding from '_utils/branding'
-import { flush, flushToHTML } from 'styled-jsx/server'
+import flush, { flushToHTML } from 'styled-jsx/server'
 
 export * from 'icon/index'